+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Improved INFO logging on auto-diskful and delete resources
+
 ## [1.32.0-rc.3] - 2025-08-26
 
 ### Added